// clients/admin-ui/src/types/api/models/ScopeRegistry.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An enumeration.
 */
export enum ScopeRegistry {
  CONFIG_READ = "config:read",
  CONFIG_UPDATE = "config:update",
  CLI_OBJECTS_CREATE = "cli-objects:create",
  CLI_OBJECTS_READ = "cli-objects:read",
  CLI_OBJECTS_UPDATE = "cli-objects:update",
  CLI_OBJECTS_DELETE = "cli-objects:delete",
  CLIENT_CREATE = "client:create",
  CLIENT_DELETE = "client:delete",
  CLIENT_READ = "client:read",
  CLIENT_UPDATE = "client:update",
  CONNECTION_CREATE_OR_UPDATE = "connection:create_or_update",
  CONNECTION_DELETE = "connection:delete",
  CONNECTION_READ = "connection:read",
  CONNECTION_AUTHORIZE = "connection:authorize",
  CONNECTION_TYPE_READ = "connection_type:read",
  CONSENT_READ = "consent:read",
  DATABASE_RESET = "database:reset",
  DATAMAP_READ = "datamap:read",
  DATASET_CREATE_OR_UPDATE = "dataset:create_or_update",
  DATASET_DELETE = "dataset:delete",
  DATASET_READ = "dataset:read",
  ENCRYPTION_EXEC = "encryption:exec",
  FIDES_TAXONOMY_UPDATE = "fides_taxonomy:update",
  GENERATE_EXEC = "generate:exec",
  MESSAGING_CREATE_OR_UPDATE = "messaging:create_or_update",
  MESSAGING_DELETE = "messaging:delete",
  MESSAGING_READ = "messaging:read",
  ORGANIZATION_CREATE = "organization:create",
  ORGANIZATION_DELETE = "organization:delete",
  ORGANIZATION_UPDATE = "organization:update",
  POLICY_CREATE_OR_UPDATE = "policy:create_or_update",
  POLICY_DELETE = "policy:delete",
  POLICY_READ = "policy:read",
  PRIVACY_REQUEST_CREATE = "privacy-request:create",
  PRIVACY_REQUEST_RESUME = "privacy-request:resume",
  PRIVACY_REQUEST_DELETE = "privacy-request:delete",
  PRIVACY_REQUEST_NOTIFICATIONS_CREATE_OR_UPDATE = "privacy-request-notifications:create_or_update",
  PRIVACY_REQUEST_NOTIFICATIONS_READ = "privacy-request-notifications:read",
  PRIVACY_REQUEST_READ = "privacy-request:read",
  PRIVACY_REQUEST_REVIEW = "privacy-request:review",
  PRIVACY_REQUEST_TRANSFER = "privacy-request:transfer",
  PRIVACY_REQUEST_UPLOAD_DATA = "privacy-request:upload_data",
  PRIVACY_REQUEST_VIEW_DATA = "privacy-request:view_data",
  RULE_CREATE_OR_UPDATE = "rule:create_or_update",
  RULE_DELETE = "rule:delete",
  RULE_READ = "rule:read",
  SAAS_CONFIG_CREATE_OR_UPDATE = "saas_config:create_or_update",
  SAAS_CONFIG_DELETE = "saas_config:delete",
  SAAS_CONFIG_READ = "saas_config:read",
  CONNECTION_INSTANTIATE = "connection:instantiate",
  SCOPE_READ = "scope:read",
  STORAGE_CREATE_OR_UPDATE = "storage:create_or_update",
  STORAGE_DELETE = "storage:delete",
  STORAGE_READ = "storage:read",
  SYSTEM_CREATE = "system:create",
  SYSTEM_DELETE = "system:delete",
  SYSTEM_UPDATE = "system:update",
  TAXONOMY_CREATE = "taxonomy:create",
  TAXONOMY_DELETE = "taxonomy:delete",
  TAXONOMY_UPDATE = "taxonomy:update",
  USER_CREATE = "user:create",
  USER_UPDATE = "user:update",
  USER_DELETE = "user:delete",
  USER_READ = "user:read",
  USER_PASSWORD_RESET = "user:password-reset",
  USER_PERMISSION_CREATE = "user-permission:create",
  USER_PERMISSION_UPDATE = "user-permission:update",
  USER_PERMISSION_READ = "user-permission:read",
  VALIDATE_EXEC = "validate:exec",
  WEBHOOK_CREATE_OR_UPDATE = "webhook:create_or_update",
  WEBHOOK_DELETE = "webhook:delete",
  WEBHOOK_READ = "webhook:read",
}